Working on AmenityConverter
Due to the Enum we have in Amenity we need to have a Converter in order for it to work properly. JPA can't read all the strings inside of the enum and instead looks for a table.
To prevent this and to make the JPA to actually read the enum correctly, we need a converter.

- Added the Converter
- Converter class comes with 2 methods, one for converting the data to a DB column, and the other to convert a DB column.